requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo>=4.11.0,<5.0.0
openai==1.3.0
python-dotenv==1.0.0
pydantic>=2.7.0

